% Generated by roxygen2 (4.0.1): do not edit by hand
\name{impactplot}
\alias{impactplot}
\title{A function to create a causal impact plot from event data.}
\usage{
impactplot(impact)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function creates the causal impact time series plot.
}
\examples{
impactplot()
}
\keyword{data}
\keyword{event}

