% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lift}
\alias{lift}
\title{A function to create a causal impact report from event data.}
\usage{
lift(causal_data, pre_period_start_date, pre_period_end_date,
  post_period_start_date, post_period_end_date, season_number, season_duration)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function creates one causal impact report.
}
\examples{
lift()
}
\keyword{data}
\keyword{event}

