% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lift_ts}
\alias{lift_ts}
\title{A function to create a causal impact report from event data in a time series format.}
\usage{
lift_ts(causal_data, data_start_date, test_start_date, season_number,
  season_duration)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function creates one causal impact report.
}
\examples{
lift_ts()
}
\keyword{data}
\keyword{event}

